+++ 03_basic_v3/src/main.js
@@ -6,6 +6,7 @@ import './style.css';
 // import App from './App04.vue';
 // import App from './App05.vue';
 // import App from './App06.vue';
+// import App from './App07.vue';
 import App from './App.vue';
 
 createApp(App).mount('#app');
